+++ src/components/domain/header/components/MenuButton.tsx
@@ -1,0 +1,33 @@
+import Image from 'next/image'
+import Link from 'next/link'
+import Button from '@/components/ui/button'
+import {
+  DropdownMenu,
+  DropdownMenuTrigger,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuItem,
+} from '@/components/ui/dropdown-menu'
+import AppPath from '@/config/appPath'
+import Assets from '@/config/assets'
+
+const MenuButton = () => {
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button className="dark:bg-white" size="icon" variant={null}>
+          <Image src={Assets.menuIcon} alt="menu" />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent>
+        <DropdownMenuGroup>
+          <DropdownMenuItem>
+            <Link href={AppPath.home()}>홈으로</Link>
+          </DropdownMenuItem>
+        </DropdownMenuGroup>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  )
+}
+
+export default MenuButton
